Delete COMaker, fold function into CommanderInfo
