+++ src/UI/GameStatsController.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 
 import CommandingOfficers.Commander;
 import CommandingOfficers.CommanderInfo;
-import CommandingOfficers.COMaker.InfoPage;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameInstance;
 import Engine.OptionSelector;
