+++ src/UI/InfoController.java
@@ -2,7 +2,7 @@ package UI;
 
 import CommandingOfficers.Commander;
 import CommandingOfficers.CommanderInfo;
-import CommandingOfficers.COMaker.InfoPage;
+import CommandingOfficers.CommanderInfo.InfoPage;
 import Engine.GameInstance;
 import Engine.IController;
